refactor: /application 페이지 전환
